Some validation on RelatedEntity and RelatedEntityType

+++ src/services/RelatedEntityTypeServices.java
@@ -5,12 +5,16 @@ import models.RelatedEntityType;
 import java.sql.*;
 import java.util.ArrayList;
 import java.util.List;
+import java.util.NoSuchElementException;
 
 public class RelatedEntityTypeServices {
 
 
 
     public static void insertRelatedEntityType( RelatedEntityType entityType)  {
+        if (entityType == null) {
+            throw new NullPointerException("entityType is null");
+        }
         String sql = "INSERT INTO RelatedEntityType (relatedEntityType) VALUES (?)";
 
         try (Connection conn = DataBaseConnection.getConnection();
@@ -64,13 +68,20 @@ public class RelatedEntityTypeServices {
     }
 
     public void updateRelatedEntityType(String oldRelatedEntityType, RelatedEntityType newEntityType) {
+        if (oldRelatedEntityType == null || newEntityType == null) {
+            throw new NullPointerException("oldRelatedEntityType or newEntityType is null");
+        }
         String sql = "UPDATE RelatedEntityType SET relatedEntityType = ? WHERE relatedEntityType = ?";
         try (Connection conn = DataBaseConnection.getConnection();
              PreparedStatement pstmt = conn.prepareStatement(sql)) {
 
             pstmt.setString(1, newEntityType.getRelatedEntityType());
             pstmt.setString(2, oldRelatedEntityType);
-            pstmt.executeUpdate();
+            int affectedrows = pstmt.executeUpdate();
+
+            if (affectedrows == 0) {
+                throw new NoSuchElementException("There is not record of the license category: " + oldRelatedEntityType);
+            }
 
         } catch (SQLException e) {
             e.printStackTrace();
@@ -78,12 +89,19 @@ public class RelatedEntityTypeServices {
     }
 
     public void deleteRelatedEntityType(String relatedEntityType) {
+        if (relatedEntityType == null) {
+            throw new NullPointerException("relatedEntityType is null");
+        }
         String sql = "DELETE FROM RelatedEntityType WHERE relatedEntityType = ?";
         try (Connection conn = DataBaseConnection.getConnection();
              PreparedStatement pstmt = conn.prepareStatement(sql)) {
 
             pstmt.setString(1, relatedEntityType);
-            pstmt.executeUpdate();
+            int affectedrows = pstmt.executeUpdate();
+
+            if (affectedrows == 0) {
+                throw new NoSuchElementException("There is not record of the license category: " + relatedEntityType);
+            }
 
         } catch (SQLException e) {
             e.printStackTrace();
